# apps/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Digitalisasi Arsip",
        "description": "REST API untuk sistem digitalisasi arsip akademik.\n\n## Autentikasi\nAPI ini menggunakan Laravel Sanctum dengan session-based authentication / Bearer Token.\n\n**Langkah autentikasi (Session):**\n1. GET `/api/csrf-cookie` untuk mendapatkan CSRF token\n2. POST `/api/auth/login` dengan credentials\n3. Sertakan cookie dan X-XSRF-TOKEN header pada setiap request\n\n## Roles\n- **Manager**: Full access ke semua fitur\n- **Uploader**: Upload, update & delete dokumen sendiri\n- **QC**: Verifikasi dokumen\n- **SBAP**: Download dokumen terverifikasi",
        "contact": {
            "name": "API Support"
        },
        "version": "4.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/prodis": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Daftar program studi aktif (UC-15)",
                "description": "Mengambil seluruh daftar program studi aktif di lingkungan fakultas.",
                "operationId": "b136b6bb0c8624cf80cfbb696a7e9517",
                "responses": {
                    "200": {
                        "description": "Daftar program studi berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProdiListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Tambah program studi baru (UC-15) - Manager only",
                "description": "Menambahkan data program studi baru ke database master data. Hanya dapat diakses oleh Manager.",
                "operationId": "ae3098bcd69ea540e2ebf64695f58e5a",
                "requestBody": {
                    "description": "Data program studi baru",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProdiRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Program studi berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProdiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden (Bukan Manager)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "description": "Reusable OpenAPI Schema Definitions.\n\nFile ini berisi semua schema yang digunakan secara berulang\ndi berbagai endpoint (request bodies, responses, dll).",
                        "type": "string",
                        "example": "The email field is invalid."
                    },
                    "errors": {
                        "description": "Object berisi field yang error dengan array pesan error",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    }
                },
                "type": "object"
            },
...
            "Prodi": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "IF"
                    },
                    "name": {
                        "type": "string",
                        "example": "Teknik Informatika"
                    },
                    "degree": {
                        "type": "string",
                        "example": "S1"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-08-18T20:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-08-18T20:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "ProdiListResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Prodi"
                        }
                    }
                },
                "type": "object"
            },
            "CreateProdiRequest": {
                "required": [
                    "code",
                    "name",
                    "degree"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "PWK"
                    },
                    "name": {
                        "type": "string",
                        "example": "Perencanaan Wilayah dan Kota"
                    },
                    "degree": {
                        "type": "string",
                        "example": "S1"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ProdiResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Program Studi berhasil ditambahkan."
                    },
                    "data": {
                        "$ref": "#/components/schemas/Prodi"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "description": "Session cookie dari Laravel Sanctum",
                "name": "digitalisasi_arsip_api_session",
                "in": "cookie"
            }
        }
    },
...
}